[backend] refactor tests to separate classes + package
