+++ backend/README.md
@@ -1,7 +1,7 @@
 ## Testing
 start test environment: `python3 src/app.py test`
 
-run tests: `python3 src/test_rest.py` or from IDE
+run tests: `python3 src/test/test_rest.py` or from IDE
 
 ## Development
 start dev environment: `python3 src/app.py dev`
